# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.AlbumRequest:
    properties:
      description:
        maxLength: 500
        type: string
      gallery_enabled:
        type: boolean
      name:
        maxLength: 100
        minLength: 1
        type: string
      serial_number:
        type: integer
    required:
    - name
    type: object
...
  github_com_leleo886_lopic_internal_error.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  services.AlbumResponse:
    properties:
      cover_image:
        type: string
      created_at:
        type: string
      description:
        type: string
      gallery_enabled:
        type: boolean
      id:
        type: integer
      image_count:
        type: integer
      name:
        type: string
      serial_number:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  services.GetAlbumImagesResponse:
    properties:
      images:
        items:
          $ref: '#/definitions/services.ImageResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  services.GetAlbumsResponse:
    properties:
      albums:
        items:
          $ref: '#/definitions/services.AlbumResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  services.ImageResponse:
    properties:
      albums:
        items:
          $ref: '#/definitions/services.AlbumResponse'
        type: array
      created_at:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_url:
        type: string
      height:
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      original_name:
        type: string
      storage_name:
        type: string
      tags:
        items:
          type: string
        type: array
      thumbnail_height:
        type: integer
      thumbnail_size:
        type: integer
      thumbnail_url:
        type: string
      thumbnail_width:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      width:
        type: integer
    type: object
...
  success.DataResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  success.SuccessResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API
  title: Lopic API
  version: 1.0.0
paths:
...
  /api/albums:
    get:
      description: 获取当前用户的相册列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.GetAlbumsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取相册列表
      tags:
      - 相册管理
    post:
      consumes:
      - application/json
      description: 创建新的相册
      parameters:
      - description: 相册信息
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/controllers.AlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.AlbumResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建相册
      tags:
      - 相册管理
  /api/albums/{id}:
    delete:
      description: 删除指定的相册（不会删除图片）
      parameters:
      - description: 相册ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除相册，只删除相册与图片的关联信息，不删除图片
      tags:
      - 相册管理
    get:
      description: 获取指定相册的详细信息
      parameters:
      - description: 相册ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.AlbumResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取相册详情
      tags:
      - 相册管理
    put:
      consumes:
      - application/json
      description: 更新指定相册的信息
      parameters:
      - description: 相册ID
        in: path
        name: id
        required: true
        type: integer
      - description: 相册信息
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/controllers.AlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.AlbumResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新相册
      tags:
      - 相册管理
  /api/albums/{id}/images:
    get:
      description: 获取指定相册中的所有图片
      parameters:
      - description: 相册ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.GetAlbumImagesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取相册中的图片
      tags:
      - 相册管理
  /api/albums/images/not-in-any:
    get:
      description: 获取所有不在任何相册中的图片
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.GetAlbumImagesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取不在任何相册中的图片
      tags:
      - 相册管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"